loja_django/requirements.txt
Django==3.2.12
Pillow==10.2.0
Stripe==8.1.0
